Make network layer, make parsing logic in presenter
Make the network layer. It belongs to the Model layer, so it's completely unaware of the presenter and the view. It communicates with the external entity via FilesNetworkLayerDelegate protocol.

Also, make the parsing logic of the received from the Google Sheets API model to the FileEntity model. If the 'parentUUID' property contains zero UUID (00000000-0000-0000-0000-000000000000) it represents the root directory.