Fix inconsistent file tree view bug
Since each FilesViewController had their own 'fileTreeViewType' and there wasn't a mechanism to syncronize their properties across all of the view controllers in the navigation controller's stack, it led to incorrect behavior of the "Tiles/table file view" button. Fix this by removing the property from the FilesViewController and providing them with a centralized storage within the navigation controller.

Also, change icons for the FileTreeViewType: we want to see the "table" icon when the screen is in tiles view, and vice versa.